Added register and comment functionality

+++ base/views.py
@@ -1,15 +1,21 @@
 from django.contrib import auth
+from django.http.response import HttpResponse
 from django.shortcuts import redirect, render
 from django.contrib.auth.models import User
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.forms import UserCreationForm
 from .models import Room, Topic
 from .forms import RoomForm
 from django.db.models import Q
 from django.contrib import messages
 from django.contrib.auth import authenticate, login, logout
 
+from base import forms
+
 
 # Create your views here.
 def loginPage(request):
+    page = 'login'
 
     if request.method == "POST":
         username = request.POST.get('username')
@@ -27,12 +33,31 @@ def loginPage(request):
         else:
             messages.error(request, 'Username or password does not exist')
 
-    context = {}
+    context = {'page': page}
     return render(request, 'base/login_register.html', context)
+
 
 def logoutUser(request):
     logout(request)
     return redirect('home')
+
+def registerPage(request):
+    form = UserCreationForm()
+
+    if request.method == "POST":
+        form = UserCreationForm(request.POST)
+        
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'An error has occured during registration')
+
+    context = {'form': form}
+    return render(request, 'base/login_register.html', context)
 
 
 def home(request):
@@ -50,15 +75,18 @@ def home(request):
     context = {'rooms': rooms, 'topics': topics, 'rooms_count': rooms_count}
     return render(request, 'base/home.html', context)
 
+
 def room(request, pk):
     room = Room.objects.get(id=pk)
-    context = {'room': room}
+    room_messages = room.message_set.all().order_by('-created')
+    context = {'room': room, 'room_messages': room_messages}
     return render(request, 'base/room.html', context)
 
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
-    
+
     if request.method == "POST":
         form = RoomForm(request.POST)
         if form.is_valid():
@@ -68,11 +96,14 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
-
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
 
+    if request.user != room.host:
+        return HttpResponse('You are not allowed!!!')
+ 
     if request.method == "POST":
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -83,8 +114,12 @@ def updateRoom(request, pk):
     return render(request, 'base/room_form.html', context)
 
 
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+    if request.user != room.host:
+        return HttpResponse('You are not allowed!!!')
+ 
     if request.method == "POST":
         room.delete()
         return redirect('home')
